refactor(Money): Money 클래스 생성 후 리팩토링
