+++ src/main/java/lotto/controller/LottoController.java
@@ -1,0 +1,60 @@
+package lotto.controller;
+
+import lotto.domain.Lotto;
+import lotto.domain.Lottos;
+import lotto.domain.Rank;
+import lotto.domain.WinningLotto;
+import lotto.view.LottoView;
+
+import java.util.ArrayList;
+import java.util.Map;
+import java.util.TreeMap;
+
+public class LottoController {
+    private static Lottos lottos;
+    private static WinningLotto winningLotto;
+    private static final ArrayList<Rank> wins = new ArrayList<>();
+    private static final Map<Rank, Integer> countByRank = new TreeMap<>();
+
+    public static void drawResult() {
+        LottoView.displayResultMessage();
+        countEachRank();
+        countByRank.forEach((rank, rankCount) ->
+            LottoView.displayResult(rank, rankCount));
+    }
+
+    private static void countEachRank() {
+        for (int i = 1; i < Rank.values().length; i++) {
+            Rank rank = Rank.values()[i];
+            int rankCount = (int) wins.stream().filter(win -> win == rank).count();
+            countByRank.put(rank, rankCount);
+        }
+    }
+
+    public void startLotto() {
+        lottos = new Lottos(LottoView.requestMoney());
+        LottoView.buyLotto(lottos.getCount());
+        LottoView.printLottos(lottos);
+        makeWinningLotto();
+    }
+
+    public void endLotto() {
+        drawLotto();
+        drawResult();
+        LottoView.displayEarningRate(Lottos.findResult(countByRank));
+    }
+
+    private void makeWinningLotto() {
+        String winningInput = LottoView.requestWinningNumber();
+        String bonusInput = LottoView.requestBonusBallNumber();
+        winningLotto = new WinningLotto(winningInput, bonusInput);
+    }
+
+    private void drawLotto() {
+        ArrayList<Lotto> lottoGroup = lottos.getLottoGroup();
+        for (Lotto lotto : lottoGroup) {
+            Rank rank = winningLotto.findRank(lotto);
+            wins.add(rank);
+        }
+    }
+}
