+++ src/main/java/lotto/domain/Lotto.java
@@ -1,8 +1,8 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
-import java.util.Set;
 import java.util.HashSet;
+import java.util.Set;
 
 public class Lotto {
     private static final String NUMBER_COUNT_ERROR = "[ERROR] 6개의 숫자를 입력해주세요";
